Added JSDocs and moved .ts files to .js files
