+++ BookAr/src/utils/image_picker.js
