+++ server/app.py
@@ -46,4 +46,4 @@ def create_app():
 
 if __name__ == "__main__":
   app = create_app()
-  app.run(host='0.0.0.0')
+  app.run(debug=True, host='192.168.0.1')
